webapp/functions/requirements.txt
firebase_functions~=0.1.0

flask
flask-restful

google-cloud-aiplatform
langchain
